Move books data to data folder
